+++ libs/qsynedit/qsynedit/highlighter/base.h
@@ -68,6 +68,7 @@ enum class HighlighterClass {
 };
 
 enum class HighlighterLanguage {
+    DecideBySuffix,
     Asssembly,
     Cpp,
     GLSL