+++ libs/redpanda_qt_utils/qt_utils/utils.h
@@ -130,7 +130,7 @@ QString extractFileDir(const QString& fileName);
 QString extractFilePath(const QString& filePath);
 QString extractAbsoluteFilePath(const QString& filePath);
 QString cleanPath(const QString& dirPath);
-QString absolutePath(const QString& dirPath, const QString& relativePath);
+QString generateAbsolutePath(const QString& dirPath, const QString& relativePath);
 QStringList absolutePaths(const QString& dirPath, const QStringList& relativePaths);
 QString escapeSpacesInString(const QString& str);